+++ test/jsontoolkit/number_test.cc
@@ -54,8 +54,8 @@ TEST(Number, string_integer_zero) {
   EXPECT_TRUE(document.is_integer());
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(0));
-  EXPECT_EQ(document, static_cast<double>(0));
+  EXPECT_EQ(document, 0);
+  EXPECT_EQ(document, 0.0);
 }
 
 TEST(Number, string_integer_positive_single_digit_integer) {
@@ -63,8 +63,8 @@ TEST(Number, string_integer_positive_single_digit_integer) {
   EXPECT_TRUE(document.is_integer());
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 3);
-  EXPECT_EQ(document, static_cast<std::int64_t>(3));
-  EXPECT_EQ(document, static_cast<double>(3));
+  EXPECT_EQ(document, 3);
+  EXPECT_EQ(document, 3.0);
 }
 
 TEST(Number, string_integer_negative_single_digit_integer) {
@@ -72,8 +72,8 @@ TEST(Number, string_integer_negative_single_digit_integer) {
   EXPECT_TRUE(document.is_integer());
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), -3);
-  EXPECT_EQ(document, static_cast<std::int64_t>(-3));
-  EXPECT_EQ(document, static_cast<double>(-3));
+  EXPECT_EQ(document, -3);
+  EXPECT_EQ(document, -3.0);
 }
 
 TEST(Number, string_integer_positive_long_integer) {
@@ -81,8 +81,8 @@ TEST(Number, string_integer_positive_long_integer) {
   EXPECT_TRUE(document.is_integer());
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 12345);
-  EXPECT_EQ(document, static_cast<std::int64_t>(12345));
-  EXPECT_EQ(document, static_cast<double>(12345));
+  EXPECT_EQ(document, 12345);
+  EXPECT_EQ(document, 12345.0);
 }
 
 TEST(Number, string_integer_negative_long_integer) {
@@ -90,8 +90,8 @@ TEST(Number, string_integer_negative_long_integer) {
   EXPECT_TRUE(document.is_integer());
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), -12345);
-  EXPECT_EQ(document, static_cast<std::int64_t>(-12345));
-  EXPECT_EQ(document, static_cast<double>(-12345));
+  EXPECT_EQ(document, -12345);
+  EXPECT_EQ(document, -12345.0);
 }
 
 TEST(Number, string_integer_minus_zero) {
@@ -99,8 +99,8 @@ TEST(Number, string_integer_minus_zero) {
   EXPECT_TRUE(document.is_integer());
   EXPECT_FALSE(document.is_real());
   EXPECT_EQ(document.to_integer(), 0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(0));
-  EXPECT_EQ(document, static_cast<double>(0));
+  EXPECT_EQ(document, 0);
+  EXPECT_EQ(document, 0.0);
 }
 
 TEST(Number, invalid_string_integer_trailing_minus) {
@@ -179,8 +179,8 @@ TEST(Number, trailing_zero_positive) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.5);
   EXPECT_EQ(document, 1.5);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(2));
+  EXPECT_FALSE(document == 1);
+  EXPECT_FALSE(document == 2);
 }
 
 TEST(Number, trailing_zero_negative) {
@@ -189,8 +189,8 @@ TEST(Number, trailing_zero_negative) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1.5);
   EXPECT_EQ(document, -1.5);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-1));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-2));
+  EXPECT_FALSE(document == -1);
+  EXPECT_FALSE(document == -2);
 }
 
 TEST(Number, single_left_digit_positive_real) {
@@ -199,8 +199,8 @@ TEST(Number, single_left_digit_positive_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.5);
   EXPECT_EQ(document, 1.5);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(2));
+  EXPECT_FALSE(document == 1);
+  EXPECT_FALSE(document == 2);
 }
 
 TEST(Number, single_left_digit_negative_real) {
@@ -209,8 +209,8 @@ TEST(Number, single_left_digit_negative_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1.5);
   EXPECT_EQ(document, -1.5);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-1));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-2));
+  EXPECT_FALSE(document == -1);
+  EXPECT_FALSE(document == -2);
 }
 
 TEST(Number, leading_decimal_zero) {
@@ -219,8 +219,8 @@ TEST(Number, leading_decimal_zero) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.0005);
   EXPECT_EQ(document, 1.0005);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(2));
+  EXPECT_FALSE(document == 1);
+  EXPECT_FALSE(document == 2);
 }
 
 TEST(Number, multi_left_digit_positive_real) {
@@ -229,8 +229,8 @@ TEST(Number, multi_left_digit_positive_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1234.5);
   EXPECT_EQ(document, 1234.5);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1234));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1235));
+  EXPECT_FALSE(document == 1234);
+  EXPECT_FALSE(document == 1235);
 }
 
 TEST(Number, multi_left_digit_negative_real) {
@@ -239,8 +239,8 @@ TEST(Number, multi_left_digit_negative_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1234.5);
   EXPECT_EQ(document, -1234.5);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-1235));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-1234));
+  EXPECT_FALSE(document == -1235);
+  EXPECT_FALSE(document == -1234);
 }
 
 TEST(Number, long_positive_real) {
@@ -249,8 +249,8 @@ TEST(Number, long_positive_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1234.56789);
   EXPECT_EQ(document, 1234.56789);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1234));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1235));
+  EXPECT_FALSE(document == 1234);
+  EXPECT_FALSE(document == 1235);
 }
 
 TEST(Number, long_negative_real) {
@@ -259,8 +259,8 @@ TEST(Number, long_negative_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1234.56789);
   EXPECT_EQ(document, -1234.56789);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-1235));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(-1234));
+  EXPECT_FALSE(document == -1235);
+  EXPECT_FALSE(document == -1234);
 }
 
 TEST(Number, multiple_sibling_periods) {
@@ -279,7 +279,7 @@ TEST(Number, single_digit_positive_real_integer) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.0);
   EXPECT_EQ(document, 1.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(1));
+  EXPECT_EQ(document, 1);
 }
 
 TEST(Number, single_digit_positive_real_integer_trailing_zero) {
@@ -288,7 +288,7 @@ TEST(Number, single_digit_positive_real_integer_trailing_zero) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), 1.0);
   EXPECT_EQ(document, 1.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(1));
+  EXPECT_EQ(document, 1);
 }
 
 TEST(Number, single_digit_negative_real_integer) {
@@ -297,7 +297,7 @@ TEST(Number, single_digit_negative_real_integer) {
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1.0);
   EXPECT_EQ(document, -1.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(-1));
+  EXPECT_EQ(document, -1);
 }
 
 TEST(Number, single_digit_negative_real_integer_trailing_zero) {
@@ -305,7 +305,7 @@ TEST(Number, single_digit_negative_real_integer_trailing_zero) {
   EXPECT_FALSE(document.is_integer());
   EXPECT_TRUE(document.is_real());
   EXPECT_EQ(document.to_real(), -1.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(-1));
+  EXPECT_EQ(document, -1);
 }
 
 TEST(Number, string_integer_with_plus) {
@@ -465,8 +465,8 @@ TEST(Number, exponential_notation_plus_after_e) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 300.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(300));
-  EXPECT_EQ(document, static_cast<double>(300));
+  EXPECT_EQ(document, 300);
+  EXPECT_EQ(document, 300.0);
 }
 
 // From https://en.wikipedia.org/wiki/Scientific_notation
@@ -476,8 +476,8 @@ TEST(Number, exponential_notation_integer_1_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 2.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(2));
-  EXPECT_EQ(document, static_cast<double>(2));
+  EXPECT_EQ(document, 2);
+  EXPECT_EQ(document, 2.0);
 }
 
 TEST(Number, exponential_notation_integer_2_upper) {
@@ -485,8 +485,8 @@ TEST(Number, exponential_notation_integer_2_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 300.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(300));
-  EXPECT_EQ(document, static_cast<double>(300));
+  EXPECT_EQ(document, 300);
+  EXPECT_EQ(document, 300.0);
 }
 
 TEST(Number, exponential_notation_integer_3_upper) {
@@ -494,8 +494,8 @@ TEST(Number, exponential_notation_integer_3_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 4321.768);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(4321));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(4322));
+  EXPECT_FALSE(document == 4321);
+  EXPECT_FALSE(document == 4322);
 }
 
 TEST(Number, exponential_notation_integer_4_upper) {
@@ -503,7 +503,7 @@ TEST(Number, exponential_notation_integer_4_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), -53000);
-  EXPECT_EQ(document, static_cast<std::int64_t>(-53000));
+  EXPECT_EQ(document, -53000);
 }
 
 TEST(Number, exponential_notation_integer_5_upper) {
@@ -511,7 +511,7 @@ TEST(Number, exponential_notation_integer_5_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 6720000000);
-  EXPECT_EQ(document, static_cast<std::int64_t>(6720000000));
+  EXPECT_EQ(document, 6720000000);
 }
 
 TEST(Number, exponential_notation_integer_6_upper) {
@@ -519,8 +519,8 @@ TEST(Number, exponential_notation_integer_6_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 0.2);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(0));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
+  EXPECT_FALSE(document == 0);
+  EXPECT_FALSE(document == 1);
 }
 
 TEST(Number, exponential_notation_integer_7_upper) {
@@ -528,7 +528,7 @@ TEST(Number, exponential_notation_integer_7_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 987);
-  EXPECT_EQ(document, static_cast<std::int64_t>(987));
+  EXPECT_EQ(document, 987);
 }
 
 TEST(Number, exponential_notation_integer_8_upper) {
@@ -536,8 +536,8 @@ TEST(Number, exponential_notation_integer_8_upper) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 0.00000000751);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(0));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
+  EXPECT_FALSE(document == 0);
+  EXPECT_FALSE(document == 1);
 }
 
 TEST(Number, exponential_notation_integer_1_lower) {
@@ -545,7 +545,7 @@ TEST(Number, exponential_notation_integer_1_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 2.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(2));
+  EXPECT_EQ(document, 2);
 }
 
 TEST(Number, exponential_notation_integer_2_lower) {
@@ -553,7 +553,7 @@ TEST(Number, exponential_notation_integer_2_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 300.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(300));
+  EXPECT_EQ(document, 300);
 }
 
 TEST(Number, exponential_notation_integer_3_lower) {
@@ -561,8 +561,8 @@ TEST(Number, exponential_notation_integer_3_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 4321.768);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(4321));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(4322));
+  EXPECT_FALSE(document == 4321);
+  EXPECT_FALSE(document == 4322);
 }
 
 TEST(Number, exponential_notation_integer_4_lower) {
@@ -570,7 +570,7 @@ TEST(Number, exponential_notation_integer_4_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), -53000);
-  EXPECT_EQ(document, static_cast<std::int64_t>(-53000));
+  EXPECT_EQ(document, -53000);
 }
 
 TEST(Number, exponential_notation_integer_5_lower) {
@@ -578,7 +578,7 @@ TEST(Number, exponential_notation_integer_5_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 6720000000);
-  EXPECT_EQ(document, static_cast<std::int64_t>(6720000000));
+  EXPECT_EQ(document, 6720000000);
 }
 
 TEST(Number, exponential_notation_integer_6_lower) {
@@ -586,8 +586,8 @@ TEST(Number, exponential_notation_integer_6_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 0.2);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(0));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
+  EXPECT_FALSE(document == 0);
+  EXPECT_FALSE(document == 1);
 }
 
 TEST(Number, exponential_notation_integer_7_lower) {
@@ -595,7 +595,7 @@ TEST(Number, exponential_notation_integer_7_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 987);
-  EXPECT_EQ(document, static_cast<std::int64_t>(987));
+  EXPECT_EQ(document, 987);
 }
 
 TEST(Number, exponential_notation_integer_8_lower) {
@@ -603,8 +603,8 @@ TEST(Number, exponential_notation_integer_8_lower) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 0.00000000751);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(0));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
+  EXPECT_FALSE(document == 0);
+  EXPECT_FALSE(document == 1);
 }
 
 TEST(Number, exponential_notation_integer_1_real) {
@@ -612,7 +612,7 @@ TEST(Number, exponential_notation_integer_1_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 2.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(2));
+  EXPECT_EQ(document, 2);
 }
 
 TEST(Number, exponential_notation_integer_2_real) {
@@ -620,7 +620,7 @@ TEST(Number, exponential_notation_integer_2_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 300.0);
-  EXPECT_EQ(document, static_cast<std::int64_t>(300));
+  EXPECT_EQ(document, 300);
 }
 
 TEST(Number, exponential_notation_integer_3_real) {
@@ -628,6 +628,6 @@ TEST(Number, exponential_notation_integer_3_real) {
   EXPECT_TRUE(document.is_real());
   EXPECT_FALSE(document.is_integer());
   EXPECT_EQ(document.to_real(), 0.2);
-  EXPECT_FALSE(document == static_cast<std::int64_t>(0));
-  EXPECT_FALSE(document == static_cast<std::int64_t>(1));
+  EXPECT_FALSE(document == 0);
+  EXPECT_FALSE(document == 1);
 }
